requirement.txt
Flask==2.3.2
python-dotenv==1.0.0
bcrypt==4.0.1
email-validator==2.0.0.post1
